requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
blurhash==1.1.4
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
coverage==7.10.3
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-csp==4.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.59.0
git-filter-repo==2.47.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
instagrapi==2.2.1
jiter==0.10.0
kiwisolver==1.4.8
Mastodon.py==2.0.1
matplotlib==3.10.3
moviepy==1.0.3
numpy==2.3.1
oauth2client==4.1.3
oauthlib==3.3.1
openai==1.97.1
packaging==25.0
pillow==11.3.0
proglog==0.1.12
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-magic-bin==0.4.14
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
ruff==0.12.8
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
tweepy==4.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
